refactor: extract fixtures to a new file
